remove About link from header and add a small footer
